fix: have local resources specify the full url
Local resources should be specified by a complete URL that way it is not
dependent on the html location. In a debug build the host should be
empty, but the path should be set to /joryvardas-ca/public-dev/ to allow
resources to still be found when testing using the builtin webserver for
WebStorm.